feat(ui): makes Chat comp more generic to fit both pages
