+++ packages/ui/app/page.tsx
@@ -1,7 +1,9 @@
 "use client";
 
+import { useEffect, useState } from "react";
+import { useCookies } from "react-cookie";
 import { HStack, VStack, Heading, Center } from "@chakra-ui/react";
-import { Chat, ChatIcon } from "../components";
+import { Chat, ChatIcon } from "@/components";
 import { ChatGPTMessage } from "@/components/ChatLine";
 
 function ChatHeader() {
@@ -33,7 +35,69 @@ export const initialMessages: ChatGPTMessage[] = [
   },
 ];
 
+const COOKIE_NAME = "nextjs-example-ai-chat-gpt3";
+
 function Home() {
+  const [messages, setMessages] = useState<ChatGPTMessage[]>(initialMessages);
+  const [loading, setLoading] = useState(false);
+  const [cookie, setCookie] = useCookies([COOKIE_NAME]);
+
+  useEffect(() => {
+    if (!cookie[COOKIE_NAME]) {
+      // generate a semi random short id
+      const randomId = Math.random().toString(36).substring(7);
+      setCookie(COOKIE_NAME, randomId);
+    }
+  }, [cookie, setCookie]);
+
+  // send message to API /api/chat endpoint
+  const sendMessage = async (message: string) => {
+    try {
+      setLoading(true);
+      const newMessage = { role: "user", content: message } as ChatGPTMessage;
+      setMessages((prevMessages) => [...prevMessages, newMessage]);
+
+      const response = await fetch(`/api/chat`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          messages: newMessage,
+          user: cookie[COOKIE_NAME],
+          chatHistory: messages,
+        }),
+      });
+
+      if (!response.ok) {
+        throw new Error(response.statusText);
+      }
+
+      const data = await response.json();
+      const noDataErrorMsg =
+        "Ops... um problema inesperado ocorreu. Aguarde alguns instantes e tente novamente.";
+
+      setMessages((prevMessages) => [
+        ...prevMessages,
+        {
+          role: "assistant",
+          content: data || noDataErrorMsg,
+        } as ChatGPTMessage,
+      ]);
+
+      setLoading(false);
+    } catch (error) {
+      console.error(error);
+      const messageWithError = {
+        role: "assistant",
+        content:
+          "Ops... Algo deu errado!\n Não estamos conseguindo carregar o chat com a IAna. Por favor, tente novamente mais tarde!",
+      } as ChatGPTMessage;
+      setMessages((prevMessages) => [...prevMessages, messageWithError]);
+      setLoading(false);
+    }
+  };
+
   return (
     <VStack spacing="10em" justify={"center"} px={4} py={6}>
       <VStack
@@ -47,7 +111,12 @@ function Home() {
         spacing={0}
       >
         <ChatHeader />
-        <Chat initialMessages={initialMessages} />
+        <Chat
+          loading={loading}
+          sendMessage={sendMessage}
+          messages={messages}
+          setMessages={setMessages}
+        />
       </VStack>
     </VStack>
   );
